Update serial_core.c to create an anonymous parent device for NI-987x
niserial987xDriver doesn't pass in a parent device pointer to uartPort when calling uart_add_one_port. This, in turn, crashes the kernel when serial_core attempts to locate the parent device pointer (since the pointer is NULL).

This PR attempts to use platform_device_register_simple to create a platform device that could be assigned as the parent device for NI-987x ports. It does this by checking if the caller is niserial987xDriver. If it is, it gets the slot number of the 987x module, and creates an appropriate platform_device for each uart port that needs to be registered.